requirements.txt
pandas
openpyxl
streamlit
openai
python-dateutil
numpy
python-dotenv
plotly